Move util method to util package from main
